updates performance page to classroom
