+++ src/components/pages/Media/View/index.tsx
@@ -10,13 +10,7 @@ const ViewMedia = ({ img, closeModal }: TViewMedia) => (
   <Modal
     size="md"
     scroll={false}
-    content={
-      <img
-        className="w-full aspect-square object-contain"
-        src={img}
-        alt="imaging"
-      />
-    }
+    content={<img className="w-full h-auto" src={img} alt="imaging" />}
     close={() => closeModal()}
   />
 );
